Add a `"binaryen-tree"` `compileMethod`, for now just a constant set inside `generator.ts`.
This is temporary while we're trying out approaches to making Wasm output.

Switch to `"moduleResolution": "nodenext"` in `tsconfig.json` so we don't have to `eval('import("binaryen")')` (avoid TS converting `import()` into a require call, so that we can use Node.js ESM from a CommonJS output file as expected).